server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["VadimK <yaralavadim@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
mypy = "^1.6.1"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
